RELEASE 1.5.7
Bug Fixes:

Stack-Related Commands: Corrected an issue from version 1.5.6 where stack-related commands were accidentally removed from input suggestions.
String and Reserved Keywords: Resolved an issue where strings identical to reserved keywords for constants were wrongly evaluated as those constants. Strings are now correctly recognized and processed when explicitly entered in quotes.

New Commands:
nip: Removes the second element from the top of the stack.
rot: Moves the third element to the top of the stack.
unrot: Moves the top element to the third position in the stack.
over: Copies the second element from the top of the stack.
roll: Moves the nth element to the top of the stack.
depth: Returns the depth of the stack, indicating how many elements are currently in the stack.